Add utility logging classes, yaml config, and logging mixin.
This WIP commit adds a core.logging module that includes a basic utility
function to get the python logger for a module, which should be used
like:

```
logger = get_logger(__name__)
```

and a LoggerMixin class which provides a `self.logger` attribute to
any class which uses it, e.g.

class Foo(LoggerMixin):
  def do_stuff:
    self.logger.info("I'm doing stuff")

The python logger configuration is loaded from a yaml file, and the
location of the logger configuration file can be overwritten by an
environment variable.