Move configurations to transformation subclasses.
Using SQLAlchemy inheritance, this commit creates specific transformation subclasses
that possess the configurations for those transformation types.  The type of the
transformation is defined by the transformation_template_id in the transformation table.

For example, there is now a ExtractTransformation type (which is a subtype of Transformation)
which now possesses the extract_configuration attribute.